* extract information about 'enclosures' from post metadata tags, pass it on for later use  * extract information about 'Image' from post metadata tags (looks like this corresponds roughly to the 'lead image' for a post).  * extract information about wordpress attachments from wp:attachment_url tags and the associated post metadata tags.  * extract information about disqus comment threads for posts, useful if you want to re-associate disqus threads later?
add a new pipeline section which downloads 'enclosures' from the wordpress site and loads them into plone as files.

Incremental improvements would be to associate enclosures with posts later via 'related items' or some similar mechanism.  Also to skip downloading any enclosures where the url either is not in a whitelist of good urls or is in a blacklist of bad urls.  Whitelist and blacklist to be set by configuration?